Extracted the cumulative code function into a hook
