+++ src/components/CodeCell.tsx
@@ -7,6 +7,7 @@ import Resizable from "./Resizable";
 import { Cell } from "../state";
 import { useActions } from "../hooks/useActions";
 import { useTypedSelector } from "../hooks/useTypedSelector";
+import { useCumulativeCode } from "../hooks/useCumulativeCode";
 
 import "./CodeCell.css";
 
@@ -18,31 +19,22 @@ const CodeCell: React.FC<CodeCellPropsType> = ({ cell }) => {
   const { updateCell, createBundle } = useActions();
 
   const bundle = useTypedSelector((state) => state.bundles[cell.id]);
-  const cumulativeCode = useTypedSelector((state) => {
-    const { data, order } = state.cells;
-    const orderedCells = order.map((id) => data[id]);
-    const cumulativeCode = [];
-    for (let c of orderedCells) {
-      if (c.type === "code") cumulativeCode.push(c.content);
-      if (c.id === cell.id) break;
-    }
-    return cumulativeCode;
-  });
+  const cumulativeCode = useCumulativeCode(cell.id);
 
   const onChange = (value: string) => updateCell(cell.id, value);
 
   useEffect(() => {
     if (!bundle) {
-      createBundle(cell.id, cumulativeCode.join("\n"));
+      createBundle(cell.id, cumulativeCode);
       return;
     }
 
     const timer = setTimeout(async () => {
-      createBundle(cell.id, cumulativeCode.join("\n"));
+      createBundle(cell.id, cumulativeCode);
     }, 1000);
 
     return () => clearTimeout(timer);
-  }, [cumulativeCode.join("\n"), cell.id]);
+  }, [cumulativeCode, cell.id]);
 
   return (
     <Resizable axis="vertical">
